refactor: rename layout component from App to Layout
