--- src/app/index.js
@@ -1,5 +1,0 @@
-import App from './App.jsx';
-
-import './app.scss';
-
-export default App;
