+++ src/layout/index.js
@@ -1,0 +1,5 @@
+import Layout from './Layout.jsx';
+
+import './Layout.scss';
+
+export default Layout;
